# docs/swagger.yaml
basePath: /v0
definitions:
  cns_errors.RainbowErrorDetailInfo:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Order:
    properties:
      amount:
        description: 单位为分
        type: integer
      app_name:
        type: string
      app_pay_notify_url:
        description: 上层应用通知url
        type: string
      app_refund_notify_url:
        description: 上层应用通知url
        type: string
      code_url:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      h5_url:
        type: string
      id:
        type: integer
      is_pay_notify_completed:
        type: boolean
      is_refund_notify_completed:
        type: boolean
      pay_notify_count:
        description: PayNotifyNextTime    *time.Time `json:"pay_notify_next_time"`
        type: integer
      qr_code_width:
        description: 二维码宽度。 只有alipay，且 trade type 为 h5 模式有效，qr pay mode 为4 时有效； 用法参考
          https://opendocs.alipay.com/apis/api_1/alipay.trade.page.pay?scene=22
        type: string
      qr_pay_mode:
        description: 支付二维码模式。 只有alipay，且 trade type 为 h5 模式有效; 用法参考 https://opendocs.alipay.com/apis/api_1/alipay.trade.page.pay?scene=22
        type: string
      refund_notify_count:
        description: RefundNotifyNextTime    *time.Time `json:"refund_notify_next_time"`
        type: integer
      refund_state:
        type: string
      return_url:
        description: 付款成功后的跳转链接。只有alipay，且 trade type 为 h5 模式有效; 用法参考 https://opendocs.alipay.com/apis/api_1/alipay.trade.page.pay?scene=22
        type: string
      time_expire:
        type: string
      trade_no:
        type: string
      trade_provider:
        type: string
      trade_state:
        type: string
      trade_type:
        type: string
      updated_at:
        type: string
    type: object
  models.OrderCore:
    properties:
      amount:
        description: 单位为分
        type: integer
      app_name:
        type: string
      code_url:
        type: string
      description:
        type: string
      h5_url:
        type: string
      qr_code_width:
        description: 二维码宽度。 只有alipay，且 trade type 为 h5 模式有效，qr pay mode 为4 时有效； 用法参考
          https://opendocs.alipay.com/apis/api_1/alipay.trade.page.pay?scene=22
        type: string
      qr_pay_mode:
        description: 支付二维码模式。 只有alipay，且 trade type 为 h5 模式有效; 用法参考 https://opendocs.alipay.com/apis/api_1/alipay.trade.page.pay?scene=22
        type: string
      refund_state:
        type: string
      return_url:
        description: 付款成功后的跳转链接。只有alipay，且 trade type 为 h5 模式有效; 用法参考 https://opendocs.alipay.com/apis/api_1/alipay.trade.page.pay?scene=22
        type: string
      time_expire:
        type: string
      trade_no:
        type: string
      trade_provider:
        type: string
      trade_state:
        type: string
      trade_type:
        type: string
    type: object
  services.MakeOrderReq:
    properties:
      amount:
        type: integer
      description:
        type: string
      notify_url:
        type: string
      qr_code_width:
        description: 二维码宽度。 只有alipay，且 trade type 为 h5 模式有效，qr pay mode 为4 时有效； 用法参考
          https://opendocs.alipay.com/apis/api_1/alipay.trade.page.pay?scene=22
        type: string
      qr_pay_mode:
        description: 支付二维码模式。 只有alipay，且 trade type 为 h5 模式有效; 用法参考 https://opendocs.alipay.com/apis/api_1/alipay.trade.page.pay?scene=22
        type: string
      return_url:
        description: 付款成功后的跳转链接。只有alipay，且 trade type 为 h5 模式有效; 用法参考 https://opendocs.alipay.com/apis/api_1/alipay.trade.page.pay?scene=22
        type: string
      time_expire:
        description: alipay 当面付无效，当面付固定过期时间为2小时
        type: integer
      trade_provider:
        enum:
        - wechat
        - alipay
        type: string
      trade_type:
        type: string
    required:
    - amount
    - description
    - time_expire
    - trade_provider
    - trade_type
    type: object
  services.MakeOrderResp:
    properties:
      code_url:
        type: string
      h5_url:
        type: string
      trade_no:
        type: string
      trade_provider:
        type: string
      trade_type:
        type: string
    type: object
  services.RefundReq:
    properties:
      notify_url:
        type: string
      reason:
        type: string
    required:
    - reason
    type: object
host: 127.0.0.1:8080
info:
  contact: {}
  description: Conflux-Pay API documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Rainbow-API
  version: "1.0"
paths:
  /orders:
    post:
      description: make order
      operationId: MakeOrder
      parameters:
      - description: make_wechat_order_req
        in: body
        name: make_ord_req
        required: true
        schema:
          $ref: '#/definitions/services.MakeOrderReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/cns_errors.RainbowErrorDetailInfo'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/cns_errors.RainbowErrorDetailInfo'
      summary: Make Order
      tags:
      - Orders
  /orders/{trade_no}:
    get:
      description: query order by trade no
      operationId: QueryOrder
      parameters:
      - description: trade no
        in: path
        name: trade_no
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: order
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/cns_errors.RainbowErrorDetailInfo'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/cns_errors.RainbowErrorDetailInfo'
      summary: query order by trade no
      tags:
      - Orders
  /orders/close/{trade_no}:
    put:
      description: close order
      operationId: Close
      parameters:
      - description: trade no
        in: path
        name: trade_no
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: order
          schema:
            $ref: '#/definitions/models.OrderCore'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/cns_errors.RainbowErrorDetailInfo'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/cns_errors.RainbowErrorDetailInfo'
      summary: close order
      tags:
      - Orders
  /orders/refresh-url/{trade_no}:
    put:
      description: refresh pay url
      operationId: RefreshPayUrl
      parameters:
      - description: trade no
        in: path
        name: trade_no
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.MakeOrderResp'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/cns_errors.RainbowErrorDetailInfo'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/cns_errors.RainbowErrorDetailInfo'
      summary: refresh pay url
      tags:
      - Orders
  /orders/refund/{trade_no}:
    put:
      description: refund pay
      operationId: Refund
      parameters:
      - description: trade no
        in: path
        name: trade_no
        required: true
        type: string
      - description: refund_req
        in: body
        name: refund_req
        required: true
        schema:
          $ref: '#/definitions/services.RefundReq'
      produces:
      - application/json
      responses:
        "200":
          description: order
          schema:
            $ref: '#/definitions/models.OrderCore'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/cns_errors.RainbowErrorDetailInfo'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/cns_errors.RainbowErrorDetailInfo'
      summary: refund pay
      tags:
      - Orders
schemes:
- http
- https
swagger: "2.0"